big decimal as currency value type

+++ src/main/java/code/challenge/currency/Currency.java
@@ -2,9 +2,11 @@ package code.challenge.currency;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.math.BigDecimal;
+
 public sealed interface Currency permits EUR, USD {
     @NotNull
-    Currency add(double amount);
+    Currency add(BigDecimal amount);
 
     @NotNull
     Currency add(@NotNull Currency other);
